Run rubocop in the parent directory that has a .rubocop.yml file.
If you are in a project that has a .rubocop.yml config file in the root
of the project, you used to have to specify an absolute path to that yml
in atom's config. This would mean you would have to change the config
each time you switch projects.

This change will find the parent directory that has the .rubocop.yml and
run rubocop from there. This allows lint-rubocop to work with mostly any
project out of the box.