Asyncify the specs
Bring in `jasmine-fix` to allow the use of `async`/`await` in the specs
and refactor them to take advantage of this. Note that the handling of
notifications was moved to the same implementation used in
`linter-eslint`. It's a more generic implementation, and properly
disposes of the subscription used to grab the notifications.