Update format for RuboCop 0.52.0
RuboCop 0.52.0 updated their format to display cop names by default in
messages. Since we were already showing the cop names this caused a few
issues in the specs.

This disables the default forced cop names in the messages, but at the
same time updates our style to match the default output of RuboCop.